# chore(mm-bot): add indexes to database