## Description
- Add indexes to database to improve performance of sorted columns.
## Results
- Queries in replica improved speed performance up to 2x.
### Before Index
**In primary:**

``` Sort  (cost=15.09..15.57 rows=190 width=389) (actual time=0.108..0.119 rows=190 loops=1)
   Sort Key: amount
   Sort Method: quicksort  Memory: 69kB
   Buffers: shared hit=6
   ->  Seq Scan on orders  (cost=0.00..7.90 rows=190 width=389) (actual time=0.011..0.035 rows=190 loops=1)
         Buffers: shared hit=6
 Planning Time: 0.059 ms
 Execution Time: 0.146 ms
(8 rows)
```

**In replica:**

```
 Sort  (cost=15.09..15.57 rows=190 width=389) (actual time=0.151..0.162 rows=190 loops=1)
   Sort Key: amount
   Sort Method: quicksort  Memory: 69kB
   Buffers: shared hit=9
   ->  Seq Scan on orders  (cost=0.00..7.90 rows=190 width=389) (actual time=0.006..0.035 rows=190 loops=1)
         Buffers: shared hit=6
 Planning:
   Buffers: shared hit=156 read=2
 Planning Time: 0.498 ms
 Execution Time: 0.212 ms
(10 rows)
```
### After Index
**In primary:**
```
 Sort  (cost=15.09..15.57 rows=190 width=389) (actual time=0.106..0.117 rows=190 loops=1)
   Sort Key: amount
   Sort Method: quicksort  Memory: 69kB
   Buffers: shared hit=6
   ->  Seq Scan on orders  (cost=0.00..7.90 rows=190 width=389) (actual time=0.007..0.031 rows=190 loops=1)
         Buffers: shared hit=6
 Planning:
   Buffers: shared hit=15 read=1
 Planning Time: 0.203 ms
 Execution Time: 0.142 ms
(10 rows)
```

**In replica:**
```
 Sort  (cost=15.09..15.57 rows=190 width=389) (actual time=0.108..0.119 rows=190 loops=1)
   Sort Key: amount
   Sort Method: quicksort  Memory: 69kB
   Buffers: shared hit=6
   ->  Seq Scan on orders  (cost=0.00..7.90 rows=190 width=389) (actual time=0.006..0.029 rows=190 loops=1)
         Buffers: shared hit=6
 Planning:
   Buffers: shared hit=21
 Planning Time: 0.171 ms
 Execution Time: 0.144 ms
(10 rows)
```